# Avoid thread divergence in intersection kernel