This commit changes the logic in the intersection kernel, specifically for the case where only a single intersection is returned, to insert the intersection into the storage only after the loop has completed, adding an implicit convergence point for the compiler to work with.