@stephenswat 2025-09-01
Before:

```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                    Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------
BM_PROPAGATION_TOY_DETECTOR_W_COV_TRANSPORT_262144_TRACKS/real_time   29755983 ns     29699359 ns           24 TracksPropagated=8.80979M/s
BM_PROPAGATION_TOY_DETECTOR_262144_TRACKS/real_time                   14458245 ns     14431920 ns           48 TracksPropagated=18.1311M/s
BM_PROPAGATION_WIRE_CHAMBER_W_COV_TRANSPORT_262144_TRACKS/real_time   30112876 ns     30056121 ns           23 TracksPropagated=8.70538M/s
BM_PROPAGATION_WIRE_CHAMBER_262144_TRACKS/real_time                   12812300 ns     12788695 ns           55 TracksPropagated=20.4603M/s
```

After:

```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                    Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------
BM_PROPAGATION_TOY_DETECTOR_W_COV_TRANSPORT_262144_TRACKS/real_time   28010923 ns     27959173 ns           25 TracksPropagated=9.35863M/s
BM_PROPAGATION_TOY_DETECTOR_262144_TRACKS/real_time                   13968447 ns     13942427 ns           50 TracksPropagated=18.7669M/s
BM_PROPAGATION_WIRE_CHAMBER_W_COV_TRANSPORT_262144_TRACKS/real_time   30161225 ns     30103998 ns           23 TracksPropagated=8.69142M/s
BM_PROPAGATION_WIRE_CHAMBER_262144_TRACKS/real_time                   12845834 ns     12822525 ns           55 TracksPropagated=20.4069M/s
```