@stephenswat 2025-09-09
Closing this, as the speedup after #1063 has become so small that it no longer matters.